Remove old table components and rename new table components
