+++ packages/web/src/components/collectibles-playlist-table/CollectiblesPlaylistTable.module.css
