+++ packages/web/src/components/collectibles-playlist-table/index.ts
@@ -1,0 +1,1 @@
+export { CollectiblesPlaylistTable } from './CollectiblesPlaylistTable'
