+++ packages/web/src/components/table/components/TableFavoriteButton.module.css
