+++ packages/web/src/components/table/components/TableFavoriteButton.tsx
