+++ packages/web/src/components/table/components/TablePlayButton.module.css
