--- packages/web/src/components/test-collectibles-playlist-table/index.ts
@@ -1,1 +1,0 @@
-export { TestCollectiblesPlaylistTable } from './TestCollectiblesPlaylistTable'
