--- packages/web/src/components/test-tracks-table/index.ts
@@ -1,1 +1,0 @@
-export { TestTracksTable } from './TestTracksTable'
