--- packages/web/src/components/tracks-table/TableDragHandle.module.css
@@ -1,18 +1,0 @@
-.tableDragHandle {
-  padding: 0px !important;
-  vertical-align: middle;
-  width: 32px !important;
-}
-
-.wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-items: center;
-  height: 16px;
-  margin-left: 16px;
-}
-
-.iconDrag path {
-  fill: currentColor;
-}
