--- packages/web/src/components/tracks-table/TableFavoriteButton.module.css
@@ -1,21 +1,0 @@
-.tableFavoriteButton.notFavorited {
-  opacity: 0;
-}
-
-:global(.ant-table-tbody > tr:hover:not(.ant-table-expanded-row))
-  .tableFavoriteButton {
-  opacity: 1;
-}
-
-.tableFavoriteButton {
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  user-select: none;
-}
-
-.icon {
-  width: 16px;
-  height: 16px;
-  position: relative;
-}
