+++ packages/web/src/components/tracks-table/index.ts
@@ -1,0 +1,1 @@
+export { TracksTable } from './TracksTable'
