+++ packages/web/src/pages/collection-page/components/desktop/CollectionPage.tsx
@@ -7,22 +7,19 @@ import {
   Variant,
   Status,
   User,
-  FeatureFlags,
   CollectionTrack,
   CollectionPageTrackRecord,
   CollectionsPageType
 } from '@audius/common'
 
-import CollectionHeader from 'components/collection/desktop/CollectionHeader'
-import Page from 'components/page/Page'
 import {
   CollectiblesPlaylistTableColumn,
-  TestCollectiblesPlaylistTable
-} from 'components/test-collectibles-playlist-table/TestCollectiblesPlaylistTable'
-import { TestTracksTable } from 'components/test-tracks-table'
-import { TracksTableColumn } from 'components/test-tracks-table/TestTracksTable'
-import TracksTable from 'components/tracks-table/TracksTable'
-import { useFlag } from 'hooks/useRemoteConfig'
+  CollectiblesPlaylistTable
+} from 'components/collectibles-playlist-table/CollectiblesPlaylistTable'
+import CollectionHeader from 'components/collection/desktop/CollectionHeader'
+import Page from 'components/page/Page'
+import { TracksTable } from 'components/tracks-table'
+import { TracksTableColumn } from 'components/tracks-table/TracksTable'
 import { computeCollectionMetadataProps } from 'pages/collection-page/store/utils'
 
 import styles from './CollectionPage.module.css'
@@ -141,7 +138,6 @@ const CollectionPage = ({
   onClickFavorites,
   onClickDescriptionExternalLink
 }: CollectionPageProps) => {
-  const { isEnabled: isNewTablesEnabled } = useFlag(FeatureFlags.NEW_TABLES)
   // TODO: Consider dynamic lineups, esp. for caching improvement.
   const [dataSource, playingIndex] =
     tracks.status === Status.SUCCESS
@@ -184,7 +180,6 @@ const CollectionPage = ({
   const isNftPlaylist = typeTitle === 'Audio NFT Playlist'
 
   const {
-    trackCount,
     isEmpty,
     lastModified,
     playlistName,
@@ -245,7 +240,7 @@ const CollectionPage = ({
   )
 
   const TableComponent = useMemo(() => {
-    return isNftPlaylist ? TestCollectiblesPlaylistTable : TestTracksTable
+    return isNftPlaylist ? CollectiblesPlaylistTable : TracksTable
   }, [isNftPlaylist])
 
   const tracksTableColumns = useMemo<
@@ -280,64 +275,34 @@ const CollectionPage = ({
           <EmptyPage isOwner={isOwner} text={customEmptyText} />
         ) : (
           <div className={styles.tableWrapper}>
-            {isNewTablesEnabled ? (
-              <TableComponent
-                // @ts-ignore
-                columns={tracksTableColumns}
-                wrapperClassName={styles.tracksTableWrapper}
-                key={playlistName}
-                loading={isNftPlaylist ? collectionLoading : tracksLoading}
-                userId={userId}
-                playing={playing}
-                playingIndex={playingIndex}
-                data={dataSource}
-                onClickRow={onClickRow}
-                onClickFavorite={onClickSave}
-                onClickTrackName={onClickTrackName}
-                onClickArtistName={onClickArtistName}
-                onClickRemove={isOwner ? onClickRemove : undefined}
-                onClickRepost={onClickRepostTrack}
-                onReorderTracks={onReorderTracks}
-                onSortTracks={onSortTracks}
-                isReorderable={
-                  userId !== null &&
-                  userId === playlistOwnerId &&
-                  allowReordering &&
-                  !isAlbum
-                }
-                removeText={`${messages.remove} ${
-                  isAlbum ? messages.type.album : messages.type.playlist
-                }`}
-              />
-            ) : (
-              <TracksTable
-                key={playlistName}
-                loading={tracksLoading}
-                loadingRowsCount={trackCount}
-                columns={columns}
-                userId={userId}
-                playing={playing}
-                playingIndex={playingIndex}
-                dataSource={dataSource}
-                allowReordering={
-                  userId !== null &&
-                  userId === playlistOwnerId &&
-                  allowReordering &&
-                  !isAlbum
-                }
-                onClickRow={onClickRow}
-                onClickFavorite={onClickSave}
-                onClickTrackName={onClickTrackName}
-                onClickArtistName={onClickArtistName}
-                onClickRepost={onClickRepostTrack}
-                onSortTracks={onSortTracks}
-                onReorderTracks={onReorderTracks}
-                onClickRemove={isOwner ? onClickRemove : null}
-                removeText={`${messages.remove} ${
-                  isAlbum ? messages.type.album : messages.type.playlist
-                }`}
-              />
-            )}
+            <TableComponent
+              // @ts-ignore
+              columns={tracksTableColumns}
+              wrapperClassName={styles.tracksTableWrapper}
+              key={playlistName}
+              loading={isNftPlaylist ? collectionLoading : tracksLoading}
+              userId={userId}
+              playing={playing}
+              playingIndex={playingIndex}
+              data={dataSource}
+              onClickRow={onClickRow}
+              onClickFavorite={onClickSave}
+              onClickTrackName={onClickTrackName}
+              onClickArtistName={onClickArtistName}
+              onClickRemove={isOwner ? onClickRemove : undefined}
+              onClickRepost={onClickRepostTrack}
+              onReorderTracks={onReorderTracks}
+              onSortTracks={onSortTracks}
+              isReorderable={
+                userId !== null &&
+                userId === playlistOwnerId &&
+                allowReordering &&
+                !isAlbum
+              }
+              removeText={`${messages.remove} ${
+                isAlbum ? messages.type.album : messages.type.playlist
+              }`}
+            />
           </div>
         )}
       </div>
