+++ packages/web/src/pages/history-page/components/desktop/HistoryPage.tsx
@@ -1,17 +1,15 @@
 import { ChangeEvent, memo, useMemo } from 'react'
 
-import { FeatureFlags, ID } from '@audius/common'
+import { ID } from '@audius/common'
 import { Button, ButtonType, IconPause, IconPlay } from '@audius/stems'
 
 import FilterInput from 'components/filter-input/FilterInput'
 import Header from 'components/header/desktop/Header'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import Page from 'components/page/Page'
-import { dateSorter } from 'components/test-table'
-import { TestTracksTable } from 'components/test-tracks-table'
+import { dateSorter } from 'components/table'
+import { TracksTable } from 'components/tracks-table'
 import EmptyTable from 'components/tracks-table/EmptyTable'
-import TracksTable from 'components/tracks-table/TracksTable'
-import { useFlag } from 'hooks/useRemoteConfig'
 
 import styles from './HistoryPage.module.css'
 
@@ -58,7 +56,6 @@ const HistoryPage = ({
   onFilterChange,
   filterText
 }: HistoryPageProps) => {
-  const { isEnabled: isNewTablesEnabled } = useFlag(FeatureFlags.NEW_TABLES)
   const tableLoading = !dataSource.every((track: any) => track.play_count > -1)
 
   const playAllButton = !loading ? (
@@ -120,8 +117,8 @@ const HistoryPage = ({
             buttonLabel='Start Listening'
             onClick={() => goToRoute('/trending')}
           />
-        ) : isNewTablesEnabled ? (
-          <TestTracksTable
+        ) : (
+          <TracksTable
             key='history'
             data={dataSource}
             userId={userId}
@@ -131,18 +128,6 @@ const HistoryPage = ({
             defaultSorter={defaultSorter}
             {...trackTableActions}
           />
-        ) : (
-          <div className={styles.tableWrapper}>
-            <TracksTable
-              userId={userId}
-              loading={tableLoading}
-              loadingRowsCount={entries.length}
-              playing={queuedAndPlaying}
-              playingIndex={playingIndex}
-              dataSource={dataSource}
-              {...trackTableActions}
-            />
-          </div>
         )}
       </div>
     </Page>
